requirements.txt
RPi.GPIO
psutil
pyserial
ublox-gps